Fix sim status display issue for esim
When esim is disabled and network settings is accessed, Use SIM
option is shown as disabled. From there, if esim is enabled Use SIM
option is not enabled. This is because getPhoneId returns default
value esim is disabled and provision status read with this
phoneid will be wrong.

Fix is to get phoneid always when querying provision status

Change-Id: I58cc7718025ddd9151019c4f756839f24b7d4520
CRs-Fixed: 2788243